Rename openliberty project
