+++ platform/openliberty/src/main/java/org/trellisldp/openliberty/WebApplication.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.app.mp.triplestore;
+package org.trellisldp.openliberty;
 
 import static java.util.Arrays.asList;
 import static org.trellisldp.app.AppUtils.printBanner;
