+++ platform/openliberty/src/main/webapp/WEB-INF/beans.xml
