+++ settings.gradle
@@ -39,10 +39,10 @@ include ':trellis-triplestore'
 include ':trellis-webac'
 include ':trellis-webdav'
 
-include ':trellis-app-mp-triplestore'
 include ':trellis-app-triplestore'
 include ':trellis-bom'
 include ':trellis-karaf'
+include ':trellis-openliberty'
 include ':trellis-osgi'
 include ':trellis-quarkus'
 include ':trellis-server'
@@ -76,10 +76,10 @@ project(':trellis-webac').projectDir = "$rootDir/components/webac" as File
 project(':trellis-webdav').projectDir = "$rootDir/components/webdav" as File
 project(':trellis-rdfa').projectDir = "$rootDir/components/rdfa" as File
 
+project(':trellis-bom').projectDir = "$rootDir/platform/bom" as File
 project(':trellis-karaf').projectDir = "$rootDir/platform/karaf" as File
+project(':trellis-openliberty').projectDir = "$rootDir/platform/openliberty" as File
 project(':trellis-osgi').projectDir = "$rootDir/platform/osgi" as File
 project(':trellis-quarkus').projectDir = "$rootDir/platform/quarkus" as File
 project(':trellis-server').projectDir = "$rootDir/platform/server" as File
-project(':trellis-app-mp-triplestore').projectDir = "$rootDir/platform/app-mp-triplestore" as File
-project(':trellis-bom').projectDir = "$rootDir/platform/bom" as File
